Do not write past end of flash (fix #71)
This patch prevents the bootloader from writing or erasing past the end
of the flash. Doing that should be harmless but apparently on STM32F3 it
is not: the beginning of the flash is erased, resulting in the
destruction of the bootloader and its configuration.

This commit adds a new platform API call; it does not implement it in
every platform. This will follow.