Reduce code size by linking with newlib-nano
This commit reduces code size enough to fix issue #76. It does so by
linking against a version of newlib optimized for code size. With this
commit enabled, the motor board platform only takes 8.7kB where it used
to overflow the 10kB limit. This fix was tested on motor-board hardware
and since it passed I also enabled it on can-io.