Separate chunk size from sector size
This commit introduces a separation between a sector (a contiguous
region of flash memory which must be erased in one block) and a chunk (a
slice of data that is sent in one datagram to be written to flash
memory).

While this distingo does not make sense for F3 platforms, where both the
chunk and sectors are 2kB , it can make a big difference in F4 ones,
where the chunk is still 2kB, but the sector can become as big as 128
kB! This means that until now a 128 kB sector would be erased 64 times,
reducing the flash lifespan and increasing the flash time greatly.

As we move on to have more and more deeply embedded F4/L4 platforms
(mostly UWB for now, and also 3 phase board in the work), a more
flexible bootloader will be required.