Set reply timeout for F4s
On the F4 the flash erase time is typically much longer than on the
F1/F3. According to the F407 datasheet, the flash erase time can be up
to 4s for a 128 kB sector. Such delays would erroneously be detected as
board not responding before bumping the timeout to 5s.

Tested on hardware (UWB beacon)