Add an erase cache to the F4 flash writer
Currently the client is based on the assumption that sectors all have a
small, constant size and that erasing them is cheap. While this was true
on F3 and on F1, this is definitely not the case on F4, where sectors
are slow to erase (up to 4 second), and whose size change greatly (from
16 kB to 128 kB).

In order to avoiding introducing complexity in the client, this patch
adds a cache to expose a uniform "erase API". The bootloader simply
takes note of when it erased a real sector and skips it when a request
wants to erase it again.

This results in large speedups for flashing the UWB beacon board, which
now "only" takes 30 seconds to flash.

Tested on hardware.