Allow different timeouts based on device size
On large devices, such as the F4s, erasing a page can take a very long
time (up to 4 seconds), and therefore we do want to have generous timeouts
as well. However, this means that the user experience is much worse in
smaller devices such as the F3s.

Since we mostly use the bootloader with small devices, the short timeout
is the default and F4 users should pass --large-pages.